tech(api): migrate /api/admin/target-profiles/{id}/badges route
