+++ api/lib/application/target-profiles/target-profile-controller.js
@@ -1,8 +1,5 @@
 import { usecases as devcompUsecases } from '../../../src/devcomp/domain/usecases/index.js';
 import * as trainingSummarySerializer from '../../../src/devcomp/infrastructure/serializers/jsonapi/training-summary-serializer.js';
-import { evaluationUsecases } from '../../../src/evaluation/domain/usecases/index.js';
-import { deserializer as badgeCreationDeserializer } from '../../../src/evaluation/infrastructure/serializers/jsonapi/badge-creation-serializer.js';
-import * as badgeSerializer from '../../../src/evaluation/infrastructure/serializers/jsonapi/badge-serializer.js';
 import * as targetProfileSerializer from '../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-serializer.js';
 import * as targetProfileSummaryForAdminSerializer from '../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-summary-for-admin-serializer.js';
 import { usecases } from '../../domain/usecases/index.js';
@@ -63,22 +60,12 @@ const findPaginatedTrainings = async function (request, h, dependencies = { trai
   return dependencies.trainingSummarySerializer.serialize(trainings, meta);
 };
 
-const createBadge = async function (request, h) {
-  const targetProfileId = request.params.id;
-  const badgeCreation = await badgeCreationDeserializer.deserialize(request.payload);
-
-  const createdBadge = await evaluationUsecases.createBadge({ targetProfileId, badgeCreation });
-
-  return h.response(badgeSerializer.serialize(createdBadge)).created();
-};
-
 const targetProfileController = {
   findPaginatedFilteredTargetProfileSummariesForAdmin,
   getTargetProfileForAdmin,
   updateTargetProfile,
   createTargetProfile,
   findPaginatedTrainings,
-  createBadge,
 };
 
 export { targetProfileController };
